committed for rectangle
